+++ MyStarcraft/Client76/TestAStar.cpp
@@ -112,7 +112,7 @@ int CTestAStar::Update( void )
 		D3DXVECTOR3 vMouse = CMouse::GetInstance()->GetPos() + m_vScroll;
 		int iIndex = int(vMouse.x / TILECX) + int(vMouse.y / TILECY) * TILEX;
 
-		if ( iIndex >= 0 && iIndex < this->m_vecTile.size() )
+		if ( iIndex >= 0 && (unsigned int)iIndex < this->m_vecTile.size() )
 		{
 			/* 시작 or 끝 위치 바꾸기.. */
 			if ( this->m_vecTile[iIndex]->byDrawID == TILE_STARTINDEX ||
@@ -141,7 +141,7 @@ int CTestAStar::Update( void )
 		D3DXVECTOR3 vMouse = CMouse::GetInstance()->GetPos();
 		int iIndex = int(vMouse.x / TILECX) + int(vMouse.y / TILECY) * TILEX;
 
-		if ( iIndex >= 0 && iIndex < this->m_vecTile.size() )
+		if ( iIndex >= 0 && (unsigned int)iIndex < this->m_vecTile.size() )
 		{
 			/* 갈 수 있는 타일 웨이 포인트로.. */
 			if ( this->m_vecTile[iIndex]->byDrawID == TILE_IDLE )
@@ -156,7 +156,7 @@ int CTestAStar::Update( void )
 		D3DXVECTOR3 vMouse = CMouse::GetInstance()->GetPos() + m_vScroll;
 		int iIndex = int(vMouse.x / TILECX) + int(vMouse.y / TILECY) * TILEX;
 
-		if ( iIndex >= 0 && iIndex < this->m_vecTile.size() )
+		if ( iIndex >= 0 && (unsigned int)iIndex < this->m_vecTile.size() )
 		{
 			if ( this->m_bMouseMoveCheck )
 			{
